Move canonicalizeAttributeName into data-interactive-utils.ts.
