+++ v3/src/data-interactive/data-interactive-utils.test.ts
@@ -1,0 +1,8 @@
+import { canonicalizeAttributeName } from "./data-interactive-utils"
+
+describe("Canonicalize Attribute Name", () => {
+  test("canonicalizeAttributeName works properly", () => {
+    expect(canonicalizeAttributeName("")).toEqual("attr")
+    expect(canonicalizeAttributeName(" fake attr(ms) ")).toEqual("fake_attr")
+  })
+})
