+++ v3/src/models/data/attribute.test.ts
@@ -2,7 +2,7 @@ import { cloneDeep } from "lodash"
 import { reaction } from "mobx"
 import { getSnapshot } from "mobx-state-tree"
 import {
-  Attribute, canonicalizeAttributeName, IAttributeSnapshot, importValueToString, isFormulaAttr,
+  Attribute, IAttributeSnapshot, importValueToString, isFormulaAttr,
   isValidFormulaAttr, kDefaultFormatStr
 } from "./attribute"
 
@@ -424,11 +424,4 @@ describe("Attribute", () => {
                 "% diff:", `${(100 * (elapsed0 - elapsed1) / elapsed0).toFixed(1)}%`)
   })
 
-})
-
-describe("Canonicalize Attribute Name", () => {
-  test("canonicalizeAttributeName works properly", () => {
-    expect(canonicalizeAttributeName("")).toEqual("attr")
-    expect(canonicalizeAttributeName(" fake attr(ms) ")).toEqual("fake_attr")
-  })
 })
