+++ v3/src/models/data/attribute.ts
@@ -327,19 +327,3 @@ export function isFormulaAttr(attr?: IAttribute): attr is IAttributeWithFormula 
 export function isValidFormulaAttr(attr?: IAttribute): attr is IAttributeWithFormula {
   return !!attr?.hasValidFormula
 }
-
-export function canonicalizeAttributeName(name: string, iCanonicalize = true) {
-  let tName = name ?? ""
-  const tReg = /\([^)]*\)$/ // Identifies parenthesized substring at end
-
-  tName = tName.trim() // Get rid of trailing white space
-  if (iCanonicalize) {
-    tName = tName.replace(tReg, '')  // Get rid of parenthesized units
-    tName = tName.replace(/\W/g, '_')  // Replace non-word characters with underscore
-  }
-  // if after all this we have an empty string replace with a default name.
-  if (tName.length === 0) {
-    tName = 'attr'
-  }
-  return tName
-}
